backend: add BlockAllocator base class (#4405)
